==> Churn/Notebooks/.ipynb_aml_checkpoints/01a-data-ingest_CSV-checkpoint2021-9-15-15-47-48Z.ipynb <==
from azureml.core import Workspace, Dataset
import numpy as np
import pandas as pd

workspace = Workspace.from_config()

dataset = Dataset.get_by_name(workspace, name='Churn')
df_pd = dataset.to_pandas_dataframe()
df_pd.head(10)

columns = ['CreditScore', 'Geography', 'Gender', 'Age',
       'Tenure', 'Balance', 'NumOfProducts', 'HasCrCard', 'IsActiveMember',
       'EstimatedSalary', 'Exited']


df_pd = df_pd[columns]
df_pd.to_csv('data/churn.csv', index=False)
