==> Churn/Notebooks/.ipynb_aml_checkpoints/01a-data-ingest_CSV-checkpoint2021-9-27-16-44-18Z.ipynb <==
import numpy as np
import pandas as pd
# import needed libraries for downloading and unzipping the file
import urllib.request
from zipfile import ZipFile
import os

# download from url
response = urllib.request.urlopen("https://raw.githubusercontent.com/Jscholtes128/Azure-AI-CustomerChurn-Hack/master/data/customerchurn.zip")
output = open('customerchurn.zip', 'wb')    # note the flag:  "wb"        
output.write(response.read())
output.close()

# unzip files
zipfile = ZipFile("customerchurn.zip")
zipfile.extract("Churn_Modelling.csv")

prep = pd.read_csv("Churn_Modelling.csv",index_col=False)


prep = prep.drop('RowNumber',axis=1)

prep.to_csv('data/turbofan.csv', index=False)

os.remove("Churn_Modelling.csv")
os.remove("customerchurn.zip")

prep.describe()

## Create a Registered Dataset from a CSV

from azureml.core.workspace import Workspace,Dataset
ws = Workspace.from_config()

dataset = Dataset.Tabular.register_pandas_dataframe(prep, ws.get_default_datastore(), "churn_from_pandas_df", show_progress=True)

#prep.register(workspace=ws,
#            name='Churn_SDK',
#            description='Churn from Notebook')
